pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "daily-briefing"
version = "2.0.0"
description = "CLI tool that generates a morning briefing — weather, calendar, git status, system health, and Kubernetes status"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.9"
authors = [{name = "wortmanb"}]
keywords = ["cli", "briefing", "morning", "dashboard", "weather", "git", "kubernetes"]
dependencies = [
    "google-api-python-client>=2.0",
    "google-auth>=2.0",
]

[project.scripts]
daily-briefing = "daily_briefing.__main__:main"

[tool.setuptools.packages.find]
include = ["daily_briefing*"]
